Add extra fields for Google
It seems that Google won't pay attention to some fields unless others
are present. So I added a bunch of extra fields to the end that are
blanks and it's happy. Also Google doesn't seem to like importing an
address in structured form, only formatted.

Also corrected date format, needs to be YYYY-MM-DD instead of
MM/DD/YYYY.